pyproject.toml
[tool.poetry]
name = "vett"
version = "0.1.0"
description = "Pluggable tool to vet open source repositories"
authors = ["Regev Golan <look4regev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
tqdm = "^4.41.0"
PyGithub = "^1.45"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
pytest = "^5.3.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
